requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai>=1.12.0
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
httpx==0.24.1
pydantic-settings==2.1.0
pytest>=7.0.0